requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
langgraph
faiss-cpu
pypdf
sentence-transformers
numpy
